requirements.txt
clickclick
requests
stups-zign
